fix: accommodation 으로 스펠링 변경
