+++ src/main/java/com/dmz/airdnd/accommodation/domain/AccommodationOption.java
@@ -1,8 +1,5 @@
-package com.dmz.airdnd.accomodation.domain;
+package com.dmz.airdnd.accommodation.domain;
 
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -19,18 +16,16 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Image {
+public class AccommodationOption {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(optional = false)
-	@JoinColumn(name = "accomodation_id", nullable = false)
-	private Accomodation accomodation;
+	@JoinColumn(name = "accomodaton_id", nullable = false)
+	private Accommodation accommodation;
 
-	@Column(nullable = false, length = 255)
-	private String url;
-
-	@Column(nullable = false)
-	private LocalDateTime createdAt;
+	@ManyToOne(optional = false)
+	@JoinColumn(name = "option_id", nullable = false)
+	private Option option;
 }
